+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- [decoder] Improve seeking accuracy by calculating exact sample offset to target position
+
 ### Fixed
 - [build] Enable building on 32-bit ARM by removing vendored protoc compiler dependency (#55)
+- [decoder] Fix channel swap that could occur after seeking
 - [http] Move AAAA record resolution timeout handling from connect to read timeout (#58)
 
 ## [v0.12.0] - 2025-01-31